+++ app/Http/routes.php
@@ -16,7 +16,8 @@ Route::get('/', function () {
 });
 
 Route::get('main','MainController@index')->name('name.main');
-Route::get('stories','MainController@stories')->name('name.stories');
+Route::get('stories','MainController@stories')->name('stories');
+Route::get('pictures','MainController@getPictures')->name('pictures');
 Route::get('painters','MainController@painters')->name('name.painters');
 Route::auth();
 Route::get('/home', 'HomeController@index');
@@ -45,13 +46,18 @@ Route::get('story','PostController@index')->name('story.index');
 Route::get('story/create','PostController@create')->name('story.create');
 Route::post('story/{story}','PostController@update')->name('story.update');
 Route::get('story/{story}','PostController@show')->name('story.show');
-Route::get('story/{story}/edit','PostController@edit')->name('story.edit');
+Route::get('story/{story}/edit','PostController@edit')->name('story.edit')->middleware('admin');
 Route::post('store','PostController@store')->name('story.store');
-Route::delete('story/{story}/','PostController@destroy')->name('story.destroy');
+Route::delete('story/{story}/','PostController@destroy')->name('story.destroy')->middleware('admin');
 
 Route::get('testrepo', 'PostController@allPosts');
 Route::auth();
 
+//Route::get('painters_pictures','MainController@painters_pictures')->name('painters_pictures');
+//Route::get('painters_pictures',function (){
+//    $user =new User;
+//
+//})
 Route::get('dashboard', [
     'middleware'=> ['web','can:admin'],
     'uses' => 'HomeController@index',
